pong_infoクラスをファイル分けした
