+++ gameplay/ponglogic/objects/__init__.py
@@ -1,0 +1,3 @@
+from .pong_info import PongInfo
+
+__all__ = ["PongInfo"]
