+++ gameplay/ponglogic/objects/components/__init__.py
@@ -1,0 +1,8 @@
+from .ball import Ball
+from .paddle import Paddle
+from .score import Score
+from .blind import Blind
+from .obstacle import Obstacle
+from .game_window import GameWindow
+
+__all__ = ["Ball", "Paddle", "Score", "Blind", "Obstacle", "GameWindow"]
