+++ gameplay/ponglogic/objects/components/blind.py
@@ -1,0 +1,6 @@
+class Blind:
+    def __init__(self):
+        self.x = 350
+        self.y = 0
+        self.width = 0
+        self.height = 0
