+++ gameplay/ponglogic/objects/components/game_window.py
@@ -1,0 +1,4 @@
+class GameWindow:
+    def __init__(self):
+        self.width = 1000
+        self.height = 600
