+++ gameplay/ponglogic/objects/components/paddle.py
@@ -1,0 +1,6 @@
+class Paddle:
+    def __init__(self):
+        self.width = 15
+        self.height = 120
+        self.left_y = 240
+        self.right_y = 240
