+++ gameplay/ponglogic/objects/components/score.py
@@ -1,0 +1,4 @@
+class Score:
+    def __init__(self):
+        self.left = 0
+        self.right = 0
